// frontend/plugins/welcome/src/api/models/EntInstitution.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SUT SA Example API Playlist Vidoe
 * This is a sample server for SUT SE 2563
 *
 * The version of the OpenAPI document: 1.0
 * Contact: support@swagger.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    EntInstitutionEdges,
    EntInstitutionEdgesFromJSON,
    EntInstitutionEdgesFromJSONTyped,
    EntInstitutionEdgesToJSON,
} from './';

/**
 * 
 * @export
 * @interface EntInstitution
 */
export interface EntInstitution {
    /**
     * 
     * @type {EntInstitutionEdges}
     * @memberof EntInstitution
     */
    edges?: EntInstitutionEdges;
    /**
     * ID of the ent.
     * @type {number}
     * @memberof EntInstitution
     */
    id?: number;
    /**
     * Institution holds the value of the "institution" field.
     * @type {string}
     * @memberof EntInstitution
     */
    institution?: string;
}

export function EntInstitutionFromJSON(json: any): EntInstitution {
    return EntInstitutionFromJSONTyped(json, false);
}

export function EntInstitutionFromJSONTyped(json: any, ignoreDiscriminator: boolean): EntInstitution {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'edges': !exists(json, 'edges') ? undefined : EntInstitutionEdgesFromJSON(json['edges']),
        'id': !exists(json, 'id') ? undefined : json['id'],
        'institution': !exists(json, 'institution') ? undefined : json['institution'],
    };
}

export function EntInstitutionToJSON(value?: EntInstitution | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'edges': EntInstitutionEdgesToJSON(value.edges),
        'id': value.id,
        'institution': value.institution,
    };
}